backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.10.4
python-dotenv==1.0.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
google-genai==1.0.0
pytest==8.3.4
httpx==0.28.1
